added .zshrc to configs
at some point I should plan out how I want to store these configs,
probably have a script that symlinks to these files with the -f flag
with the users choice of the default location (specified in a file) or a
custom location. for terminal config stuff probably store that in a seperate
file and link it in to the .zshrc and .bashrc dynamically.